+++ src/sentry_options.h
@@ -29,9 +29,8 @@ struct sentry_attachment_s {
  * the sentry internals.
  */
 typedef struct sentry_options_s {
-    char *raw_dsn;
-    sentry_dsn_t dsn;
     double sample_rate;
+    sentry_dsn_t *dsn;
     char *release;
     char *environment;
     char *dist;
@@ -56,7 +55,14 @@ typedef struct sentry_options_s {
     /* everything from here on down are options which are stored here but
        not exposed through the options API */
     struct sentry_backend_s *backend;
-    sentry_user_consent_t user_consent;
+
+    long user_consent;
+    long refcount;
 } sentry_options_t;
+
+/**
+ * Increments the reference count and returns the options.
+ */
+sentry_options_t *sentry__options_incref(sentry_options_t *options);
 
 #endif