+++ src/sentry_transport.h
@@ -3,6 +3,7 @@
 
 #include "sentry_boot.h"
 
+typedef struct sentry_dsn_s sentry_dsn_t;
 typedef struct sentry_run_s sentry_run_t;
 typedef struct sentry_rate_limiter_s sentry_rate_limiter_t;
 
@@ -73,7 +74,8 @@ typedef struct sentry_prepared_http_request_s {
  * rate limited.
  */
 sentry_prepared_http_request_t *sentry__prepare_http_request(
-    sentry_envelope_t *envelope, const sentry_rate_limiter_t *rl);
+    sentry_envelope_t *envelope, const sentry_dsn_t *dsn,
+    const sentry_rate_limiter_t *rl);
 
 /**
  * Free a previously allocated HTTP request.